requirements.txt
## Web Application
Django==1.8.2
django-dotenv==1.3.0
django-autoslug==1.8.0
django-grappelli==2.6.5
django-model-utils==2.2
django-jsonfield==0.9.13
djangorestframework==3.1.3
whitenoise==2.0
psycopg2==2.6.1

## User input from Markdown
Markdown==2.5.2
bleach==1.4.1
html5lib==0.999

## Natural Language Processing
nltk==3.0.0

## Social Authentication
python-social-auth==0.2.10
python-openid==2.2.5
oauthlib==0.7.2
requests-oauthlib==0.5.0
PyJWT==1.3.0

## Other Dependencies
six==1.9.0
python-dateutil==2.4.2
PyYAML==3.11
requests==2.7.0

## Testing
nose==1.3.7
mock==1.0.1
coverage==3.7.1
coveralls==0.5
